pyproject.toml
[project]
name = "ai-agent-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cloud-sql-python-connector>=1.18.3",
    "fastapi[standard]>=0.115.14",
    "functions-framework>=3.8.3",
    "google-api-python-client>=2.176.0",
    "googledriver>=0.1.7",
    "gspread>=6.2.1",
    "line-bot-sdk>=3.17.1",
    "oauth2client>=4.1.3",
    "openai>=1.93.0",
    "pandas>=2.3.1",
    "pg8000>=1.31.2",
    "pymupdf4llm>=0.0.26",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.41",
]
